requirements.txt
tensorflow==2.5.0
streamlit==0.84.0
numpy==1.19.5
pillow==8.2.0